Fixed float error

+++ plot.py
@@ -31,10 +31,10 @@ def genPlotUn(result, fname, dirName, udStream, numFeatures):
     ax.set_ylabel('Reads')
 
     plt.figure(figsize=(15, 10))
-    plt.plot(x[(len(x) - (udStream/4*3) - 1):len(x) - 1],
-             y[(len(x) - (udStream/4*3) - 1):len(x) - 1], linestyle='-.', color='red')
+    plt.plot(x[(len(x) - int(udStream/4*3) - 1):len(x) - 1],
+             y[(len(x) - int(udStream/4*3) - 1):len(x) - 1], linestyle='-.', color='red')
              
-    plt.plot(x[(udStream/4*3):udStream - 1], y[0:(udStream/4*3) - 1], linestyle='-.', color='red')
+    plt.plot(x[int(udStream/4*3):udStream - 1], y[0:int(udStream/4*3) - 1], linestyle='-.', color='red')
     
     plt.plot(x[udStream - 1:(len(x) - udStream - 1)],
              y[udStream - 1:(len(x) - udStream - 1)], color='black')
